Fixing a strange search box CSS issue
Recently I've spotted various instances where custom skins define a height
for the search box and for whatever reason, it nowadays renders badly when
it didn't use to. I've no idea what's the underlying cause, but removing
the height definition from the appropriate search box rule always seems to
fix it, so why not?

Change-Id: Ic0f40ee235009844c0912183530ac8d290b80b87